requirements.txt
opencv-python
aiohttp
motor
python-dotenv

# Development and testing dependencies
pytest
pytest-aiohttp